Move PP session feedback to AppointmentsController
This reuses uses the function that was originally moved across from
the `ServiceProviderReferralsController` I think it's fairly readable,
and nice to have some reusable code, but I hope it's not too hacked
together to cover params possibly not being present in the URL.

I've decided to keep the legacy `/action-plan/...` endpoint in here but
not include it in the re-used code as we're hoping to remove it fairly
soon once usage of it drops off.
