+++ server/routes/probationPractitionerReferrals/probationPractitionerReferralsController.ts
@@ -201,79 +201,6 @@ export default class ProbationPractitionerReferralsController {
     ControllerUtils.renderWithLayout(res, view, expandedServiceUser)
   }
 
-  async viewSubmittedPostSessionFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { referralId, sessionNumber } = req.params
-
-    const referral = await this.interventionsService.getSentReferral(accessToken, referralId)
-
-    if (referral.actionPlanId === null) {
-      throw createError(500, `action plan does not exist on this referral '${req.params.id}'`, {
-        userMessage: 'No action plan exists for this referral',
-      })
-    }
-
-    const currentAppointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      referral.actionPlanId,
-      Number(sessionNumber)
-    )
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(currentAppointment)
-
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    if (!referral.assignedTo) {
-      throw new Error('Referral has not yet been assigned to a caseworker')
-    }
-
-    const presenter = new SubmittedFeedbackPresenter(
-      currentAppointment,
-      new AppointmentSummary(currentAppointment, referral.assignedTo, deliusOfficeLocation),
-      serviceUser,
-      'probation-practitioner',
-      referral.id,
-      null
-    )
-    const view = new SubmittedFeedbackView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  // This is left to keep links in old emails still working - we'll monitor the endpoint and remove it when usage drops off.
-  async viewLegacySubmittedPostSessionFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-
-    const actionPlan = await this.interventionsService.getActionPlan(accessToken, actionPlanId)
-    const referral = await this.interventionsService.getSentReferral(accessToken, actionPlan.referralId)
-    const currentAppointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      actionPlanId,
-      Number(sessionNumber)
-    )
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(currentAppointment)
-
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    if (!referral.assignedTo) {
-      throw new Error('Referral has not yet been assigned to a caseworker')
-    }
-
-    const presenter = new SubmittedFeedbackPresenter(
-      currentAppointment,
-      new AppointmentSummary(currentAppointment, referral.assignedTo, deliusOfficeLocation),
-      serviceUser,
-      'probation-practitioner',
-      referral.id,
-      null
-    )
-    const view = new SubmittedFeedbackView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
   async viewSubmittedPostAssessmentFeedback(req: Request, res: Response): Promise<void> {
     const { user } = res.locals
     const { accessToken } = user.token
